+++ plugins/message-bridge/src/protocol/downstream/index.ts
@@ -1,4 +1,0 @@
-export * from './SupportedDownstreamMessages.js';
-export * from './DownstreamMessageTypes.js';
-export * from './GatewayBusinessMessageAdapter.js';
-export * from './InvalidInvokeToolErrorResponder.js';